feat: Complete project
Much of this project was based on the MPC lesson completed prior to doing the project, especially the MPC code.

This commit adds functionality to the MPC class. Some improvements made are:
- Multipliers to change cost functions, which affects the smoothness of the driving and ability to adhere to the ground truth trajectory
- A polynomial of order 3 is used to calculate the predicted trajectory
- The predicted trajectory is saved and returned so it can be displayed by the sim
- Points are transformed from vehicle coordinates to world coordinates
- Predicted and ground truth paths are displayed in the sim

Some basic errors also had to be corrected, such as the lack of `std::` scope resolution operators anywhere.