feat: Account for 100ms actuator delay in simulation
Previously, the delay had been accounted for by simply keeping the vehicle speed low so it was never an issue. Now, the delay is added to each component of the state vector using the state update equations. This allows the car to drive at a much higher speed.

`rubric_answers.md` was also completed in accordance with the Udacity rubric.